Revert "Use long double to compute edge distance." (r3334). Simple double in HCoordinate made #350 pass anyway.
git-svn-id: http://svn.osgeo.org/geos/trunk@3340 5242fede-7e19-0410-aef8-94bd7d2200fb

+++ include/geos/geomgraph/EdgeIntersection.h
@@ -49,13 +49,13 @@ public:
 	geom::Coordinate coord;
 
 	// the edge distance of this point along the containing line segment
-	long double dist;
+	double dist;
 
 	// the index of the containing line segment in the parent edge
 	int segmentIndex;
 
 	EdgeIntersection(const geom::Coordinate& newCoord,
-	                 int newSegmentIndex, long double newDist)
+	                 int newSegmentIndex, double newDist)
 	  :
 	  coord(newCoord),
 	  dist(newDist),
@@ -74,7 +74,7 @@ public:
 
 	int getSegmentIndex() const { return segmentIndex; }
 
-	long double getDistance() const { return dist; }
+	double getDistance() const { return dist; }
 
 };
